refactor: split analysis/mcp/dispatch.py into focused modules
